Recognize singletons patterns via functions
Summary:
Compared with the previous diff D40181325 that can recognize the simple singleton pattern (in which the value of a global variable is directly checked, e.g. `Foo::$bar is null`), this diff handles some more complex cases, in which functions like `idx` or `C\contains_key` are used to access the elements of global collections.

Limitations:
- If the value of `C\contains_key` is stored in some local variable (instead of directly used in the conditional), then we lose track of the nullable variables.
- There could be more functions like `C\contains_key` that checks the existence of some elements in the collection or `idx` that returns the given collection's element, but are not recognized by the current algorithm. If we encounter such functions, we will need to generalize our algorithm by using **an explicit list of such functions**.

Reviewed By: minjang

Differential Revision: D40420507

fbshipit-source-id: fd9ab68bcae827edc281aa332da55262b2b9ae39