Generate entities for all parameters and returns
Summary:
We were avoiding generating entities for paramters and returns that cannot possibly be dicts.

In local mode, this leads to results that are sneakily not local. For example, a `dict` could be returned without invalidation by embedding the `dict` inside a `vec` as a `vec` cannot possibly be a `dict`.

This was happening because we invalidate the parameter and return entities, but if there is no entity generated, then there is nothing to invalidate.

This diff changes this behaviour by not being parsimonious.

The diff also causes a regression in the test suite because of the imprecise handling of collections tracked in T131709581.

Reviewed By: mheiber

Differential Revision: D39854947

fbshipit-source-id: 25bd1c2bcac7250e1706f5fe012a740044ee0581