Allow refinements in `classname<+T>`
Summary:
`classname` is covariant in its type parameter, and
a refinement of a type is a subtype of the unrefined type
(e.g., `Box with { type T = ... }` is a subtype of `Box`),
therefore we should allow, for instance:

  classname<Box with { type T = int }>

Instead of banning this, apply the instantiability TAST check
recursively on the LHS of refinement, which effectively treats the above as:

  classname<Box>

and is consistent with how HackC sees refinements in positions that don't expose TypeStructure.
This is future-proof in case we later allow refinements of non-classish types
(rather than outright allowing _any_ refinement).

Without this, the usability suffers as one needs to do
introduce a new type parameter with a bound, which also
gives false positive [lint warnings](https://www.internalfb.com/diff/D39230855?fbid=526180459507394&dest_fbid=4947380592030982) that the new type is unused:

  function alternative<Tbox as Box with { type T = int }>(
    classname<Tbox> $arg,
  ): void {}

Reviewed By: hgoldstein

Differential Revision: D40568669

fbshipit-source-id: 6112529a93e7185ba12861fa75e6e0d501ec48f0