Refactor PeekingManager to delegate to appropriate ReadCallback API
Summary:
`PeekingManager` currently uses the `getReadBuffer` and `readDataAvailable` option for reading data from the underlying socket. However, the wrapped `ReadCallback` might support movable buffers. Once the `PeekingManager` completes its peeking and unregisters itself, the wrapped `ReadCallback` might be called with the `readBufferAvailable`, putting it in an inconsistent state. Use whatever API the `ReadCallback` supports, preferring the movable buffer option.

This currently happens with Rocket's `Parser` if Fizz is being used and if no ALPN is provided. Cpp2Worker doesn't find a client ALPN, so installs a peeking manager. The peeking manager reads some bytes with the Rocket frame, decides to use the `RocketRoutingHandler`, and hands the `RocketServerConnection` the bytes looked at with the `getReadBuffer` method. When more data is ready to be read, Fizz uses the `readBufferAvailable` API in `Parser` and the code chokes, reading something inappropriate.

Reviewed By: maalbash

Differential Revision: D39604214

fbshipit-source-id: 17b475692fa3ad30dec048ca7dc9ef55fdcc957d