Simplify EqFunc to EqCls
Summary:
EqFunc is used to check the callee against the profiled call target. If we are
checking against a method in non-overridden class, compare its class instead.
This allows us to sink the Func load into the unlikely branch and avoid an
unncessary memory load.

Added assert in getMethod() reveals out of bound read in prof-data-serialize
logic, so fix that.

Reviewed By: ricklavoie

Differential Revision: D40652510

fbshipit-source-id: aa0aaecd80cade2a34893fd54403ce56f6c731c5