Optimize folly::to<> for 128-bit integers to string
Summary:
The current version does one 128-bit division for each digit outside of the 64-bit range, which is extremely slow.
Instead we can rewrite the integer in base-10^19 (requiring at most 2 128-bit divisions), so that each component fits in a 64-bit integer and we can use the existing optimized functions.

Reviewed By: philippv, luciang

Differential Revision: D40675286

fbshipit-source-id: df0616ed518d9ac68c1956c2dd585d2029ce4fa0