Fix refinement for like-types
Summary:
Previously, when trying to refine a like-type with is/as, the type parameter environment could get set to inconsistent because at was attempting to enforce the hint type to be a sub-type of `dynamic`, which it won't be. This is a soundness bug, and it is fixed here by adopting a simpler refinment rule for like types.

If we are refining `~C<t>` to `D<_>`, then we can form the intersection `~C<t> & D<T#1>`, distribute to `(dynamic & D<T#1>) | (C<t> & D<T#1>)`. We can then use the super-type (by throwing away part of the intersection) `(dynamic) | (C<t> & D<T#1>)` which is just `~(C<t> & D<T#1>)` and we can simplify the intersection as in the non-like-type case.

The above strategy is too imprecise when the hint type is enforceable, e.g., a primitive type or a class type w/o generic. In particular, if we are refining `~t` to `string`, it is not helpful to end up with `~string`. In this case, we do not throw anything away and keep the full `(dynamic & D) | (C<t> & D)`. In the common case where `D <: C<t>` this whole type just simplifies to `D`.

Reviewed By: andrewjkennedy

Differential Revision: D40508582

fbshipit-source-id: 83f3aea36ec7b183dd6a1ddc736b7f67d2d8903f