Be liberal when producing a non-collection entity
Summary:
In this diff we exploit types attached to expressions to reduce invalidation of legitimate results that arise from our simple-minded treatment of collections.

In short, because collections and their contents are identified by the same entity, when `$d['key']` flows into an inter-procedural entity (such as being passed as an argument to a function), we invalidate `$d` because we cannot distinguish `$d` from `$d['key']`. However, we can use the type of `$d['key']` to identify whether it is disjoint from any of the collections and if so stop tracking it.

Reviewed By: mheiber

Differential Revision: D40008463

fbshipit-source-id: 6fe8453a3d48ae797caed514e9fd320cf2a91e0f