Be liberal when interacting with global entities with permissive types
Summary:
When a result flows into an inter-procedural entity in local mode, we would invalidate that result as the local mode wouldn't generate directives to update non-local definitions.

However, this is too conservative in the case the property or the function parameter has a `mixed` type. In that case, there is no update to other definitions needed.

This diff makes the analysis more liberal by selectively omitting the invalidation by inspecting types.

It currently applies to property assignments and function calls.

Reviewed By: mheiber

Differential Revision: D39929753

fbshipit-source-id: 8da2c9ecd9810fffb4fbbd09588a95dd1ce1b14f