Replace \unpack with \ord in SecureRandom\string implementation
Summary:
Functionally there's nothing different between the new and old implementation.

Unpack is totally unnecessary when dealing with a one-byte string like this. `ord` is sufficient. This should save a significant amount of gCPU given that f_unpack is around .12% gCPU and a lot of it is in the random_string function according to Strobelight.

Differential Revision: D40447250

fbshipit-source-id: 6c7544fcfc3100b94793b0a458178b27fd6dbe65