propagate inheritance: fix two bugs
Summary:
### What

Fix the two bugs described in the linked tasks, as shown in updates to snapshot tests.

### How

> There is a lot I don't yet understand about the analysis, so parts of this explanation could be wrong or misleading

Both before and after this change, there are two main functions to deal with inheritance of constants:
- add_const_inheritance_constraints: adds subset constraints for conts
- propagate_const_inheritance_constraints: does the substitution

This diff changes *when* we call these functions, which increases generality.

Before:
- before: piggybacking on where constants are used and not fully searching up the inheritance hierarchy to find definitions

After:
- whenever there is a class constant, resolve where the class is defined and propagate/close

## Future work

Both before and after this change, we reimplement part of the semantics of Hack's OO features. It will be less work and easier to get correctness if we ask the decl_provider where constants are defined, rather than figuring it out ourselves.

Reviewed By: madgen

Differential Revision: D40355451

fbshipit-source-id: d1c1b8fd73a97c41d789d6b995b18b701904e3f3