Example of incorrect interprocedural solve: handling hierarchy with multiple overrides of constants
Summary:
See the snapshot test for bug T134628593

I hypothesize that the root cause is that when we add subset constraints between constants we don't inspect all ancestor classes: instead we only generate subset constraints for the current class and at most one parent class. The problem is inherent to the design of `Hips_solver.constr_list_and_string_of`.

As the snapshot test shows, this problem happens even for keys that are accessed, which makes this bug distinct from T134611333, though perhaps a redesign could fix both issues.

Reviewed By: madgen

Differential Revision: D40298199

fbshipit-source-id: 71435a2d5542f34fb89dc2aacd3db05e518f057a