Example of incorrect interprocedural solve: should leave this as a dict
Summary:
Here's a case where if we convert to shape then `hh` will error.

The runtime behavior is to fatal.

Is it OK to ignore these cases, with the expectation that `hh` will catch them? Or will it help for codemodding if our analysis is smart enough to not convert to shapes here?

Reviewed By: madgen

Differential Revision: D40181614

fbshipit-source-id: 7076ae7e2dfa6ff0b1a8d7484a826fdbe87d0e83