Rewrite AwaitAllWaitHandle::fromContainer in hack
Summary: I came here to fix the fact that the native implementation of AwaitAllWaitHandle::fromContainer was the only place manually using the mangled names of native static methods (e.g. c_AwaitAllWaitHandle_ns_fromVec) but then I realized this builtin doesn't need to be native, we can easily rewrite it in hack. Do so.

Reviewed By: jano

Differential Revision: D40503654

fbshipit-source-id: 7e83ed51196136ce19eb9794ac10c2f3fd8c35c4