Remove buf_into_value
Summary:
Unlike other ocamlrep_slab entrypoints which extract slabs from byte slices, this function does not attempt to verify that the slab is valid before using it. This places a lot of safety obligations on the caller which are undocumented:

1. The byte slice must contain a valid slab (i.e., contains a root value offset pointing to a valid value in the slab data region, every offset value in the slab data region is in bounds, etc)
2. The given byte slice must have the same alignment as the byte slice into which the slab was written
3. The slab must have been written into the byte slice by a machine with the same native endianness as the reading machine

Other entrypoints do at least the bare-minimum validation of `Slab::check_initialized` (which looks for the magic number written to initialized slabs and verifies that the root value offset is in bounds), which helps mitigate all three safety obligations (a valid slab at the wrong alignment or endianness will not appear to contain the magic number, and an invalid slab is unlikely to contain the magic number).

It looks like the caller of `buf_into_value` is not doing anything to ensure that (2) is upheld. It happens to work because jemalloc seems to allocate backing storage for `Vec<u8>` with 8-byte alignment, but AFAIK there is no firm guarantee of this. Unless there is such a guarantee, we ought to avoid relying on it.

We can do that by properly serializing the OwnedSlab instead of treating the slice of memory it's stored in as if it's a serialization format. This means that we need not worry about safety obligations (2) and (3) above; they are both addressed by the implementations of `OwnedSlab::{serialize,deserialize}`.

Perhaps the author of `buf_into_value` didn't do this because they weren't interested in paying the cost of `OwnedSlab::deserialize` (which does a relatively expensive check of the slab data region in order to uphold safety obligation (1) above). This diff adds an unsafe `deserialize_unchecked` method for this use case (which uses `Slab::check_initialized` as a sanity check) and documents the safety obligation.

Reviewed By: shayne-fletcher

Differential Revision: D40690609

fbshipit-source-id: ce2dc4b03d9062e14be4d1cf8958d48d6810dceb