Remove OpaqueValue
Summary:
The existence of SlabAllocator (removed in the previous diff) constrained us to return OpaqueValue from to_ocamlrep rather than Value. Since SlabAllocator values used offsets instead of pointers, we needed to prevent consumers from dereferencing them in safe code (which Value allows, and *must* allow for FromOcamlRep to be implementable in safe code). OpaqueValue was the way we achieved that.

Now that SlabAllocator is gone, we no longer allocate OpaqueValues using offsets, and there is no need to stop them from being dereferenced. We can simplify the ocamlrep API by removing it.

Reviewed By: shayne-fletcher

Differential Revision: D40742456

fbshipit-source-id: f5a7a4f87db4aa6d68a8fe94c44ef0e9fe229983