Remove SlabAllocator
Summary:
SlabAllocator is a perf optimization: instead of invoking ToOcamlRep to allocate a value in an ocamlrep::Arena, then copying the resulting value into a contiguous vector, we construct the contiguous vector during the ToOcamlRep conversion.

Unfortunately, it places a lot of constraints on the signature of ToOcamlRep: we cannot allocate pointer-values using actual memory addresses, since the contiguous vector may need to be resized during the conversion. So instead, we use offsets into that vector, and fix them up with a rebase operation afterward.

This requires the to_ocamlrep method to return OpaqueValue rather than Value. A non-immediate Value is a pointer and can be dereferenced in safe code. A non-immediate OpaqueValue may either be a pointer or an offset. Since we cannot dereference an offset (without access to the address of the underlying vector which it is an offset into), we must provide in the ocamlrep::Allocator trait a `set_field` method (which either dereferences pointers, or indexes into the underlying vector). If it weren't for SlabAllocator, this wouldn't be necessary--BlockBuilder could implement IndexMut (and did, before SlabAllocator).

This is too much complexity for a perf optimization in a feature we intend to remove anyway. Remove SlabAllocator so that we can simplify ocamlrep's API.

Reviewed By: shayne-fletcher

Differential Revision: D40734434

fbshipit-source-id: 68a4f0ee11d21cc92148aeb9b303d593a580a0c4