Don't run nix-flake-update in forked repository
The purpose of the GitHub Action workflow `nix-flake-update` is to build HHVM with latest nixpkgs dependencies in a pull request, so we could know if there is any security updates or breaking changes in nix dependencies.

However it cannot trigger CI in forked repositories because the deploy SSH private key used is for `facebook/hhvm` only. Therefore it is useless for forked repositories.

This PR disables it in forked repositories.