Disallow override of async by non-async
Summary:
There is a consensus on the Hack Pack workchat that it's a mistake for Hack to allow `async` methods to be overridden by non-async methods that return `Awaitable`. This pattern is a problem for Sound Dynamic pessimisation, as it is an example of an override that doesn't preserve enforcement (as an `async` method returning `Awaitable<int>` enforces `int` but a non-async method only enforces `Awaitable`).

Let's outlaw this in Hack. A separate diff for www asyncifies the ~220 instances of this pattern in www.

Reviewed By: vassilmladenov

Differential Revision: D38940712

fbshipit-source-id: 64b0e3d63feeff4a945486307e3ddee15984ef9d