Don't run nix-flake-update in forked repositories (#9278)
Summary:
The purpose of the GitHub Action workflow `nix-flake-update` is to build HHVM with latest nixpkgs dependencies in a pull request, so we could know if there is any security updates or breaking changes in nix dependencies.

However it cannot be triggered in forked repositories because the deploy SSH private key used is for `facebook/hhvm` only. Therefore it is useless for forked repositories.

This PR disables it in forked repositories.

Pull Request resolved: https://github.com/facebook/hhvm/pull/9278

Test Plan: The GitHub Action workflow `nix-flake-update`  should not be triggered in a forked repository any more once the forked repository sync its default branch to a commit including this PR.

Reviewed By: aorenste

Differential Revision:
D40778535

LaMa Project: L1088355

Pulled By: Atry

fbshipit-source-id: 96b9fd9223513adeb232bdde608e42b225b5ca7c