Workaround to find HPHP::Func::s_funcVec symbol used by walkstk
Summary:
GDB is failing to find the `HPHP::Func::s_funcVec` symbol, which is used by `walkstk` when looking up jitted frame information. I don't know the exact cause yet, but running `ptype`, `ptype`, and `maint flush symbol-cache` (exactly in that order, that many times) is fixing it.

I verified this symbol is in the DWARF info.

Reviewed By: jtwarren

Differential Revision: D40821276

fbshipit-source-id: fc75fa36e30ec95b485e4983cb01118fb2974103