change data structure to store hash table
Summary:
In  D40147010 (https://github.com/facebook/hhvm/commit/c35808a4f7755141bbb05b43e6adc94ac65bddc5), I used an ordered set to store the table of AST hashes. This degraded the indexer performance significantly. I did some experimentation and It wasn't the table lookup, but seemed directly related to the footprint of the data structure.

I replaced the set with a flat array of concatenated hashes (~50MB and this solves the performance issue).

## Before this diff

https://www.internalfb.com/intern/skycastle/run/103582791439311237/actions

It takes 50min to index a shard.

## After the diff

https://www.internalfb.com/intern/skycastle/run/58546795165625035/actions

It takes 10-15min (this uses full-init as it is an ephemeral fbpkg, it'll take less than 10min once it's landed and uses the load-native-state init).

Reviewed By: donsbot

Differential Revision: D40814071

fbshipit-source-id: f2cefeac39a781b27c74115af74d909bf575f102