Remove setThreadManager calls from smcproxy 3/N
Summary: We are deprecating ThreadManager. This is part of a stack of resources that adjusts smcproxy so that it can run in either ThreadManager or Resource Pools mode. smcproxy is a complicated example because it shares the same ThreadManager amongst three services. ResourcePools allows you to share the executor pool (the threads that do the work) amongst multiple services with some work but not the queues.

Reviewed By: ivzhh

Differential Revision: D38502280

fbshipit-source-id: 7baa3514e629b22c353e1aef3d9adaad605395d9