Change StringInterner to use internal mutability.
Summary:
It's pretty common to need to be able to add a new string to a StringInterner and having to pass around an owned one is pretty onerous.

Instead use internal mutability so we can pass around a borrowed StringInterner instead.

This doesn't seem to have any perf effect on a "verify ir" run of www.

Reviewed By: edwinsmith

Differential Revision: D40806105

fbshipit-source-id: 178c76affdbd08d0ab019763b63866f17ad8f6a7