Don't use the presence of $this in the frame when backtracing
Summary:
The compact backtrace code was looking at whether the frame has $this to determine whether to put type '::' or type '->' in the frame (i.e. whether the method being called is a static or instance method). This is unreliable as we may not have $this in the frame for a variety of reasons (e.g. await tail call optimization and frames with locals already decrefed that we're looking at on exit in profilers).

Have the code look at whether the method was declared static instead, same as the non-compact code path does. Add a unit test that exercises this code (and would previously fail in repo mode due to async tail call optimization).

Also fix expectations in a test that was collecting backtraces in a profiler on function exit and which was seeing frames with locals already decrefed (so it never had $this).

Reviewed By: ricklavoie

Differential Revision: D40873064

fbshipit-source-id: d9e879c875e0152e9e5621afad253a4d7a44e1e6