Eliminate multiple applications of `naming_elaborate_namespaces_endo` visitor
Summary:
A number of the functions exposed by the Naming module first apply the `naming_elaborate_namespaces_endo` visitor to the syntactic element before applying other naming transformations.

However, these functions are also called by `program_` which itself applies the same visitor. As a result, we will be needlessly visiting certain parts of the AST multiple times.

This diff breaks the functions into helpers, which drives the non-visitor elaboration and is called from `program`, and the exposed functions which apply the visitor then call the helper

Reviewed By: CatherineGasnier

Differential Revision: D40474770

fbshipit-source-id: 19b141cf276f3729fe1678db8994b3b72b9f135b