Add a heuristic for selecting disjuncts during subtyping
Summary: During subtyping, we sometimes end up with a constraint t <: ?#0 which becomes t <: #0 || nonnull <: #0 if we don't know that t <: nonnull. From there which disjunct that type inference actually uses is dependent on the ordering. This adds a heuristic to prefer the t <: #0 side over the nonnull side regardless of the ordering.

Reviewed By: andrewjkennedy

Differential Revision: D40805236

fbshipit-source-id: cc3ac83809b31ed6bc9dab0128a23e1ad4d6ac87