Distinguish Singleton and Caching pattern
Summary: This diff distinguishes Caching pattern from Singleton pattern. Essentially, both patterns writes to (or say, initializes) a global variable when its value is null or does not exist. Yet, it is desired to distinguish if the written global variable is a collection or not: if the assignment directly writes to a collection (e.g. `Foo::$dic[$key] = $value`), then we assume it's a caching; otherwise, it is a singleton (e.g. `Foo::$bar = $x` or `(Foo::$dic[$key])->prop = $y`).

Reviewed By: lavenzg

Differential Revision: D40872670

fbshipit-source-id: 0ec2e66767ef3d1fe241749e4de9607c88222a3e