Output the function type for memoized functions
Summary:
Previously, when we raise a global read caused by memoized functions/methods, we only print out the return type of memoized function. As required by BA (i.e. there is the need to learn the parameter types of memoized functions, especially if a vc of type IViewerContext is used as the parameter), this diff prints out the whole function type of memoized function, which will be stored in "data_type" column of our DB table.

To do: create a separate table to store the parameter types.

Reviewed By: lavenzg

Differential Revision: D40880563

fbshipit-source-id: 73d1c566383fc36674a03bdcb8c9bd8ed7be6b4c