Add t_release_state.
Summary:
Splitting the release state from if that state is allowed in a given context, simplifying the enums, so we can actually compare them. This way we can enforce thrift.NoTesting for thrift.Release without forcing that on other use cases of the annotations *and* do 'range' checks on the release state.

Actual enforcement to come in subsiquent diffs.

Reviewed By: joshkehn

Differential Revision: D40760869

fbshipit-source-id: ae2c376ef0b62fa1877b40fdba7d4ebcb023672e