Fix off-by-one error in dead FIXME removal
Summary:
FIXME removal through `hh --remove-dead-fixmes` was broken due to an off-by-one error.

The error is that because we explicitly go one more than the position end, if the character after an inline FIXME comment is significant (non-whitespace) we corrupt the source code and lead to parsing as well as a slew of other (best-effort) errors.

This error was introduced in D9990364 (https://github.com/facebook/hhvm/commit/1a91cf918eeec482138c412abdd913aa9ac73ae6) to fix an off-by-one error, but it has no tests and no context, so I don't know what it fixes. The tests don't fail as a result of this change and it fixes the broken behaviour.

Reviewed By: jthemphill

Differential Revision: D40908435

fbshipit-source-id: b63aa40b4c78c081068cf59e1b7952f40ee33160