Output builtin decls for each translation unit
Summary:
Having builtins written once in the end is suboptimal because we'd have to wait
for the processing of all files to complete before we read builtins and start
the translation in infer.

With this diff, each unit in the output is self-sufficient. An alternative
might be to output builtins as a common section *at the beginning* -- we can
consider it as an optimization later on.

Reviewed By: geralt-encore

Differential Revision: D40637287

fbshipit-source-id: 65632835c2d75b7882d35b9e9481adc83fc3aabf