Limit concurrency when downloading all data in hhbbc
Summary:
When we reach make_local() and want to download all data
generated with extern-worker, we're quite aggressive. However, this
can result in excessive memory usage. Limit our max concurrency (live
coroutines) when doing the downloading. This does not affect the time
taken in a meaningful way, but does cap the memory usage.

Reviewed By: edwinsmith

Differential Revision: D40924528

fbshipit-source-id: 79510a8ade97db68b9833be1afb9a02b44ea99b6