Split NamedEntity table into namedTypes + namedFunctions
Summary:
Background: the NamedEntity table is a case-insensitive map from symbol name
to a NamedEntity. Hack/PHP toplevel types and functions are in logically separate
namespaces, so each NamedEntity has enough state to represent information about
a function, a type, and various other bits that only apply to types or functions but not
both.

In practice, there are only a tiny number of NamedEntity that contain *both*,
i.e. a function and type with the same (case insensitive) name.

The fact that they're all in the same table presents a problem for logging
as we migrate to case-sensitive symbols. case-collisions report false positives,
for example this is perfectly legal (both before and after we have case sensitive
names), and should never be logged as a case insensitive match:

  class X {}
  function x() {}

We don't want to suppress *all* logging for the named entity table, and selectively
logging by filtering out false-positives such as the example, in code paths doing
queries, turns out to be a mess.

This diff explores splitting the named entity table into two tables. Here are some
discussion points:
 * splitting the tables is orthogonal to whether names are case sensitive or insensitive
   since classes and functions are already separate namespaces, and lookups are
   always syntax-directed.
 * if we like this design, we can further refine NamedEntity into NamedClass +
   NamedFunction, since no entry would ever contain both.
 * Splitting the tables may allow us to migrate to case-sensitive toplevel types and
   functions separately instead of all-at-once.

Reviewed By: ricklavoie

Differential Revision: D37717566

fbshipit-source-id: c33d8eaa05cbf8c9464a1f81bdacbccaab770ca8