Type __toString() methods as returning `string`
Summary: D40877889 adds a placeholder return type to each function which lacks one. In doing so, I discovered that it's a Hack naming violation for a function named `__toString()` to be typed as returning anything other than a `string`. So I'm typing every `__toString()` function as returning a `string` in this diff.

Reviewed By: hgoldstein

Differential Revision: D40921440

fbshipit-source-id: c443841f5e05f905b09b5143c3207f48b893d29a