Remove dependency between BoxedValuePtr and op::compare
Summary:
BoxedValuePtr doesn't use Tag, thus T will always use `<` as less and `==` as equal, in which case there is no need to add op::compare as dependency.

In addition, in the future we will support thrift struct in op::compare. Without this diff we will have circular dependency, since

op::compare depends on Field Ref, which depends on BoxedValuePtr, which depends on op::compare

Reviewed By: iahs

Differential Revision: D40843004

fbshipit-source-id: 0a2810c2c18f3ccfb4253697b79b58ef0be49f93