Type SplFileObject::current()
Summary:
`SplFileObject` is used to iterate over the lines of a file.

So `SplFileObject::current()` returns the current line of a file, which is a `string`: https://www.php.net/manual/en/splfileobject.current.php

Unless.

Unless you set the `SplFileObject::READ_CSV` flag, in which case it *parses* the current line of the file as a CSV and returns each column as an entry in a vec.

Or unless it's reached the end of the file, in which case `SplFileObject::current()` returns `false`.

Cool.

This means a few things:

1. We should type `SplFileObject` as an `Iterator<string>`. Typing it as a `RecursiveIterator<SplFileObject>` makes no sense, because its `Tv` is not `SplFileObject`.
2. We can't trust `SplFileObject::current()` to return any particular value. It will *probably* return a string, unless you've set a specific flag or called `current()` on an exhausted iterator. But it won't always return a string.

This is why we have like-types! I'm marking `SplFileObject` to implement `Iterator<~string>`, indicating that it probably returns a string but that this type might be wrong. By marking *untrusted* types this way, we open the door to have types we can actually *trust* later on down the line.

Reviewed By: vassilmladenov

Differential Revision: D40922421

fbshipit-source-id: d125df8ef5c48a658c8f16e6ea6a3da0c6e3b71d