Get rid of a few more (4K) unused RATE metrics
Summary:
We emit a large number (30K on slb.regional) of RATE metrics. Mostly, these come from pool and VIPs.
These metrics duplicate the SUM ones. RATE is just SUM/60 and can be simulated on fetch. I removed a few more of the dashboard references recently with D40388925.

This change removes some of the RATE metrics in cases where they are not used and duplicated with SUM.
This is about 4000, or 4% of the metrics emitted in slb.regional (canary P545114069 vs control P545114071).

I also annotated some others that I wanted to remove, but could not because they are being used. These usages should be converted to SUM prior to removing RATE.

Reviewed By: jalopezsilva

Differential Revision:
D40407188

LaMa Project: L1141030

fbshipit-source-id: e74d4a7ad9373170031c711d711349715ba8fa4e