Merge all the cmake build test exclusions into one file
Summary: This avoids the need to update 2 files when excluding a test from both hhvm-ubuntu jobs ang github actions, which was error prone. The downside is that we exclude more tests than we strictly need to on hhvm-ubuntu, but we think that's an acceptable tradeoff to make the update process more reliable.

Differential Revision: D40989622

fbshipit-source-id: 09c2b2f56d76a1df4132e8febd503cc58ec0bdde