Thread through dependency state in `unit-cache.cpp`
Summary: The in memory unit cache attempts to re-use units if they exist in the cache and have a SHA1 hash identical to that of the on-disk file. This isn't dependency aware, so it's broken for DDB. What we can do here is thread through _why_ we believe the cache is stale. I've changed `isChanged` to return an enum indicating whether we think there's no change (where we returned `false` before), whether the change is considered due to a `stat` difference, or some other reason (difference in repo options, we're "stressing" the unit cache).

Reviewed By: edwinsmith

Differential Revision: D40815495

fbshipit-source-id: e9b0d437a0c3af9c064fb37fd6ff72d1c0f5e9e9