move a bunch of hhbc pub re-exports from sub-modules to lib.rs
Summary: A number of publicly used hhbc definitions were re-exported from various places in the IR causing clients to have to dig into the IR crates to import them.  Instead they should all be re-exported from the top-level so if we move them to be IR-native definitions we can change it without having to update clients.

Reviewed By: edwinsmith

Differential Revision: D40839485

fbshipit-source-id: a88bcbabc22b94ac39eb92e288ccfbea5253e2cd