tracer for changes to objects in request heap
Summary:
This diff introduces a way to monitor refcount changes of certain objects.

When allocating a HeapObject (often an object of a specific class), we could use `HeapObjectSanitizer` instead of `MemoryManager`, to allocate the object on a separate  page of memory. To track any writes to the word containing the refcount, the page containing the refcount is set to read-only using `mprotect(PROT_READ)`. Any subsequent attempt to write to the page will immediately cause a SIGSEGV (segmentation fault).

We write a custom signal handler that deals with SIGSEGV without actually crashing. The handler does 4 things.
1. Check whether SIGSEGV was triggered while accessing the protected pages, and if so, call `mprotect(PROT_READ | PROT_WRITE)` on the page, to allow the write to go through.
2. If the write changes a monitored word (usually the refcount), record the stack trace of the access and print out the current value of refcount. This allows us to track all refcount changes on the object.
3. Parsing the instruction stream, and change the *next* instruction int3. This allows the *current* instruction to finish, and then immediately causes a SIGTRAP, so that we can change the page to read-only again (in handler of SIGTRAP) and catch more future writes.
4. Manipulate the %RIP register to go back in the instruction stream, to retry the SIGSEGV-inducing instruction and do proper writes to the monitored page.

Upon SIGTRAP, the signal handler uses `mprotect` to mark pages as read-only when appropriate.

The dynamic change in instruction stream and page protection while multiple threads are running is racy. Efforts have been made to make sure a thread restores instruction/page protection to the original state after it makes changes. Correctness of the implementation relies on the fact that the traced objects are only accessed in one thread.

Reviewed By: jano

Differential Revision: D40840339

fbshipit-source-id: 3bf26590e4e3355f7e1e5ae7d3ceee66ffab96ed