OSS TOC structure
Summary:
New TOC structure for the OSS content. With this diff, we can:
* define a custom sidebars.js file where we can control the ordering of topics listed in the side bar menu
* create custom categories with auto-generated cards for subpages and folders within the category

It also has a referenced to the website/fb/sidebars.js file that is currently set to auto-gen the files/folders under /doc/fb directory. A separate diff to customize TOC for our Meta Development TOC will follow...

Reviewed By: Alfus, vitaut

Differential Revision: D40878665

fbshipit-source-id: 5261d05e40e1a5107a0d74c0fbe429c542a83fa5