Merge all the cmake build test exclusions into one file (#9291)
Summary:
Pull Request resolved: https://github.com/facebook/hhvm/pull/9291

This avoids the need to update 2 files when excluding a test from both hhvm-ubuntu jobs ang github actions, which was error prone. The downside is that we exclude more tests than we strictly need to on hhvm-ubuntu, but we think that's an acceptable tradeoff to make the update process more reliable.

Reviewed By: Atry

Differential Revision: D40989622

fbshipit-source-id: abb7afda674bf31f4f700df8c0610325fe639257