Give nearly every untyped function in Hack's `.hhi` the type `HH\FIXME\MISSING_RETURN_TYPE`
Summary:
If D40920785 (https://github.com/facebook/hhvm/commit/0c5aa56b300692e7a2ddb92c2953feef642ce1f4) is accepted, I'll change this diff to make each function return `HH\FIXME\MISSING_RETURN_TYPE` instead.

This is part of an effort to remove the `Tany` type from WWW, and by extension, from Hack. The goal is for Hack to distinguish between **trusted** types (types that have been proven from first principle. If a variable isn't within its trusted type, then it's a bug in the typechecker, runtime, or both) and **untrusted** types (types which the user might have lied about through something like an `HH_FIXME`). Right now, users can use `HH_FIXME` comments nearly anywhere, so *every* type is untrusted.

When a function doesn't have a return type, the Hack typechecker implicitly generates a type called `Tany`. This type acts as both a top type (every type is a `Tany`) and as a bottom type (a `Tany` is a subtype of every other type). This setup is very flexible, but Hack will never be sound as long as it has a concept of a `Tany` type.

This diff codemods each function which doesn't have a return type to return `HH\FIXME\TANY_MARKER<dynamic>`. This type acts like a `Tany` today, but in the future it will act as the sound `dynamic` type.

Reviewed By: vassilmladenov

Differential Revision: D40877889

fbshipit-source-id: 3baedc4cfdcd7769ed21c8eb82a35bdb4864b3d9